+++ p2f_client/harm_location.py
@@ -29,7 +29,7 @@ class harm_location:
         if health_check(self.base_url):
             r = requests.post(self.hdl_url, 
                               data=self.p2fclient.jswa("new_location", new_location.model_dump_json(exclude_unset=True)),
-                            headers={"Content-Type": "application/json"})
+                            headers=self.p2fclient.base_headers)
             return HARM_Location(**r.json())
     def list_harm_locations(self,
                             bounding_box: Optional[HARM_Bounding_Box]=None,
@@ -70,8 +70,8 @@ class harm_location:
         params = {x: y for x, y in params.items() if y != None}
         if health_check(self.base_url):
             r = requests.get(self.hdl_url,
-                            params=params, data=self.p2fclient.jswa(),
-                            headers={"Content-Type": "application/json"})
+                            params=params,
+                            headers=self.p2fclient.base_headers)
             return [HARM_Location(**x) for x in r.json()]
     def get_harm_location(self, location_identifier: UUID) -> HARM_Location:
         """Retrieve an individual HARM Location by its location identifier
@@ -82,8 +82,8 @@ class harm_location:
         :rtype: p2f_pydantic.harm_data_metadata.HARM_Location
         """
         if health_check(self.base_url):
-            r = requests.get(self.hdl_url/str(location_identifier), data=self.p2fclient.jswa(),
-                            headers={"Content-Type": "application/json"})
+            r = requests.get(self.hdl_url/str(location_identifier),
+                            headers=self.p2fclient.base_headers)
             return HARM_Location(**r.json())
     def delete_harm_location(self, location_identifier: UUID):
         """Delete a location from the API
@@ -92,8 +92,8 @@ class harm_location:
         :type location_identifier: UUID
         """
         if health_check(self.base_url):
-            r = requests.delete(self.hdl_url/str(location_identifier), data=self.p2fclient.jswa(),
-                            headers={"Content-Type": "application/json"})
+            r = requests.delete(self.hdl_url/str(location_identifier),
+                            headers=self.p2fclient.base_headers)
     def assign_location_to_record(self, location_identifier: UUID, record_hash: str):
         """Assign a location to a record hash
 
@@ -108,8 +108,8 @@ class harm_location:
         assign_url.args["location_identifier"] = str(location_identifier)
         assign_url.args["record_hash"] = record_hash
         if health_check(self.base_url):
-            r = requests.post(assign_url, data=self.p2fclient.jswa(),
-                            headers={"Content-Type": "application/json"})
+            r = requests.post(assign_url,
+                            headers=self.p2fclient.base_headers)
     def remove_location_from_record(self, location_identifier: UUID, record_hash: str):
         """Remove a location assigned to a record hash
 
@@ -124,5 +124,5 @@ class harm_location:
         remove_url.args["location_identifier"] = str(location_identifier)
         remove_url.args["record_hash"] = record_hash
         if health_check(self.base_url):
-            r = requests.post(remove_url, data=self.p2fclient.jswa(),
-                            headers={"Content-Type": "application/json"})
+            r = requests.post(remove_url,
+                            headers=self.p2fclient.base_headers)