+++ p2f_client/harm_reference.py
@@ -27,7 +27,7 @@ class harm_reference:
         """
         if health_check(self.base_url):
             r = requests.post(self.hr_url, data=self.p2fclient.jswa("new_reference", new_reference.model_dump_json(exclude_unset=True)),
-                            headers={"Content-Type": "application/json"})
+                            headers=self.p2fclient.base_headers)
             if r.ok:
                 return (HARM_Reference(**r.json()))
     def list_harm_references(self) -> List[HARM_Reference]:
@@ -39,8 +39,8 @@ class harm_reference:
         :rtype: List[p2f_pydantic.harm_reference.HARM_Reference]
         """
         if health_check(self.base_url):
-            r = requests.get(self.hr_url, data=self.p2fclient.jswa(),
-                            headers={"Content-Type": "application/json"})
+            r = requests.get(self.hr_url,
+                            headers=self.p2fclient.base_headers)
             if r.ok:
                 return [HARM_Reference(**x) for x in r.json()]
             else:
@@ -54,8 +54,8 @@ class harm_reference:
         :rtype: p2f_pydantic.harm_reference.HARM_Reference
         """
         if health_check(self.base_url):
-            r = requests.get(self.hr_url / reference_id, data=self.p2fclient.jswa(),
-                            headers={"Content-Type": "application/json"})
+            r = requests.get(self.hr_url / reference_id,
+                            headers=self.p2fclient.base_headers)
             if r.ok:
                 return HARM_Reference(**r.json())
     def delete_harm_reference(self, reference_id: UUID):
@@ -65,8 +65,8 @@ class harm_reference:
         :type reference_id: UUID
         """
         if health_check(self.base_url):
-            r = requests.delete(self, reference_id, data=self.p2fclient.jswa(),
-                            headers={"Content-Type": "application/json"})
+            r = requests.delete(self, reference_id,
+                            headers=self.p2fclient.base_headers)
     def assign_harm_reference(self, 
                               reference_id: UUID, 
                               record_hash: str):
@@ -81,8 +81,8 @@ class harm_reference:
         assign_url.args["reference_id"] = reference_id
         assign_url.args["record_hash"] = record_hash
         if health_check(self.base_url):
-            r = requests.post(assign_url, data=self.p2fclient.jswa(),
-                            headers={"Content-Type": "application/json"})
+            r = requests.post(assign_url,
+                            headers=self.p2fclient.base_headers)
     def remove_harm_reference(self, 
                               reference_id: UUID, 
                               record_hash: str):
@@ -97,5 +97,5 @@ class harm_reference:
         remove_url.args["reference_id"] = reference_id
         remove_url.args["record_hash"] = record_hash
         if health_check(self.base_url):
-            r = requests.delete(remove_url, data=self.p2fclient.jswa(),
-                            headers={"Content-Type": "application/json"})
+            r = requests.delete(remove_url,
+                            headers=self.p2fclient.base_headers)